Adding JWT token expiry timestamp validation
The JWT tokens are created which expire after TokenExpiryDuration.
Added check to validate the received JWT token timestamp to ensure
that the token was created within TokenExpiryDuration (default configured
value: 15 minutes) prior to receiving the request. This ensures that no
malformed token with indefinite expiry timestamp are accepted by the server.

Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>